Tweak discover script
