+++ src/pages/realms/musicalCubes/sounds/index.ts
@@ -1,1 +1,3 @@
-export const tracks = { Daydreamer : ["./assets/sounds/musicalCube/tracks/Daydreamer/melody/AHA_SV_70_brass_stack_loop_daydreamer_Gm.wav","./assets/sounds/musicalCube/tracks/Daydreamer/melody/AHA_SV_70_guitar_stack_loop_daydreamer_Gm.wav","./assets/sounds/musicalCube/tracks/Daydreamer/melody/AHA_SV_70_keys_loop_daydreamer_Gm.wav","./assets/sounds/musicalCube/tracks/Daydreamer/melody/AHA_SV_70_vocal_loop_daydreamer_Gm.wav"],InTheCity : ["./assets/sounds/musicalCube/tracks/InTheCity/melody/AHA_SV_60_brass_stack_loop_in_the_city_6-8_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/melody/AHA_SV_60_guitar_stack_loop_in_the_city_6-8_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/melody/AHA_SV_60_piano_loop_in_the_city_6-8_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/melody/AHA_SV_60_vocal_loop_in_the_city_6-8_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/one_shots/AHA_SV_brass_stack_one_shot_stab_choo_choo_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/one_shots/AHA_SV_melody_stack_one_shot_stab_bump_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/one_shots/SNS_SF_808_sub_comp_Am.wav"] }
+const tracks = { Daydreamer : ["./assets/sounds/musicalCube/tracks/Daydreamer/melody/AHA_SV_70_brass_stack_loop_daydreamer_Gm.wav","./assets/sounds/musicalCube/tracks/Daydreamer/melody/AHA_SV_70_guitar_stack_loop_daydreamer_Gm.wav","./assets/sounds/musicalCube/tracks/Daydreamer/melody/AHA_SV_70_keys_loop_daydreamer_Gm.wav","./assets/sounds/musicalCube/tracks/Daydreamer/melody/AHA_SV_70_vocal_loop_daydreamer_Gm.wav"],InTheCity : ["./assets/sounds/musicalCube/tracks/InTheCity/melody/AHA_SV_60_brass_stack_loop_in_the_city_6-8_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/melody/AHA_SV_60_guitar_stack_loop_in_the_city_6-8_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/melody/AHA_SV_60_piano_loop_in_the_city_6-8_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/melody/AHA_SV_60_vocal_loop_in_the_city_6-8_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/one_shots/AHA_SV_brass_stack_one_shot_stab_choo_choo_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/one_shots/AHA_SV_melody_stack_one_shot_stab_bump_Am.wav","./assets/sounds/musicalCube/tracks/InTheCity/one_shots/SNS_SF_808_sub_comp_Am.wav"] } 
+
+export default tracks;
